refactor: service→server backflow buckets 2-4 (Part of #3037)
Relocate misplaced domain transport types and cluster infrastructure out
of the `server` layer back into `services`, reducing the
`service_server_backflow` audit metric from 61 to 44 (-17).

Bucket 4 — dispatch delivery transport types:
- Move `DispatchMessagePostError`, `DispatchMessagePostErrorKind`,
  `DispatchMessagePostOutcome`, `DispatchNotifyDeliveryResult`, and
  `ReviewFollowupKind` from `server/dto/dispatches.rs` (which mixed them
  with genuine serde response DTOs) into
  `services::dispatches::discord_delivery::transport`, beside the delivery
  logic that produces/consumes them. The delivery-result roundtrip tests
  move with the types. The dispatch facade re-exports them, so route-layer
  callers are unchanged. This removes the dto→service→server→service
  ping-pong import.

Bucket 3 — cluster node-registry / session-routing infrastructure:
- Relocate `server/cluster.rs` -> `services/cluster/node_registry.rs` and
  `server/cluster_session_routing.rs` -> `services/cluster/session_routing.rs`.
  Both are pure cluster coordination (config + db + serde) with no
  route/axum dependency, so they belong beside the rest of the cluster
  services. `server::cluster` and `server::cluster_session_routing` become
  thin re-export facades so route-layer consumers keep their paths.
  Service-side callers now import the canonical
  `services::cluster::node_registry::*` / `::session_routing::*` paths.

Mechanical relocation only — no behavior change. Baseline regenerated and
change-surfaces.md / ARCHITECTURE.md synced for the +1 LoC wrap in
runtime_bootstrap.rs from the longer cluster path.

Deferred (logic-requiring or relay hot files, left for follow-up):
- Bucket 2 review-decision service extraction + loopback-HTTP removal
  (touches turn_bridge/completion_guard.rs — relay hot file, off-limits).
- dispatched_sessions.rs auto_queue::activate loopback + ws emits.
- ws event bus + monitoring store relocation (entangled with AppState in
  server/state.rs; one consumer is the relay hot file tmux_watcher.rs).

Gates: cargo fmt --check clean, cargo check --workspace --all-features
--all-targets clean, cargo build green, cluster/dispatch/dto unit tests
green (47 passed), ./scripts/ci-script-checks.sh exit 0.
