+++ src/services/cluster/mod.rs
@@ -11,10 +11,20 @@
 pub(crate) mod intake_router_hook;
 pub(crate) mod intake_routing;
 pub(crate) mod intake_worker;
+/// Worker-node registry + capability routing infrastructure. Relocated from
+/// `server::cluster` (#3037 bucket 3): it is pure cluster coordination
+/// (config + db + serde) with no route/axum dependency, so it belongs beside
+/// the rest of the cluster services. `server::cluster` re-exports it for the
+/// route layer.
+pub(crate) mod node_registry;
 pub mod registry_adapter_sink;
 pub mod relay_producer_registry;
 pub mod session_discovery;
 pub mod session_matcher;
 pub mod session_registry;
+/// Session owner routing helpers. Relocated from
+/// `server::cluster_session_routing` (#3037 bucket 3); `server::cluster_session_routing`
+/// re-exports it for the route layer.
+pub(crate) mod session_routing;
 pub mod stream_relay;
 pub mod watcher_supervisor;
