+++ src/services/dispatched_sessions.rs
@@ -266,7 +266,7 @@ pub async fn list_dispatched_sessions(
     if let Some(pool) = state.pg_pool_ref() {
         return match dispatched_sessions_db::list_dispatched_sessions_pg(pool, include_all).await {
             Ok(mut sessions) => {
-                let worker_nodes = match crate::server::cluster::list_worker_nodes(
+                let worker_nodes = match crate::services::cluster::node_registry::list_worker_nodes(
                     pool,
                     state.config.cluster.lease_ttl_secs.max(1),
                 )
@@ -280,7 +280,7 @@ pub async fn list_dispatched_sessions(
                         Vec::new()
                     }
                 };
-                crate::server::cluster_session_routing::enrich_session_owner_routing(
+                crate::services::cluster::session_routing::enrich_session_owner_routing(
                     &mut sessions,
                     state.cluster_instance_id.as_deref(),
                     &worker_nodes,
